requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pygments==2.16.1
